requirements.txt
celery==5.1.2
Django==4.0
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
requests==2.26.0
django-extensions==3.1.3
djangorestframework_simplejwt==5.0.0
django-rest-passwordreset==1.2.1
drf-spectacular==0.21.0
drf-spectacular[sidecar]==0.21.0
psycopg2==2.9.2
pymysql==1.0.2
django-redis-cache==3.0.0
redis==3.5.3
django-jazzmin==2.4.8